+++ src/routes/blog/+page.ts
@@ -1,6 +1,6 @@
 import type { PageLoad } from './$types';
 import type { ArticleMetadata, TagWithCount, ArticleThumbnailImgFmts } from '$lib/scripts/types';
-import { getTags } from '$lib/scripts/tag/utils';
+import { getTags } from '$lib/btpc/scripts/utils';
 
 export const load: PageLoad = async ({
 	url,
